requirements.txt
feedparser
datetime
python-dotenv
jinja2
